aider: Implemented a fix to remove todos from the list when the delete key is pressed after tabbing to a todo item.
# Aider chat conversation:

USER: Pressing the delete key does nothing after I have tabbed to a todo item.

ASSISTANT: To fix the issue where pressing the delete key does nothing after tabbing to a todo item, we need to modify the `handleDelete` function in `my-app/components/TodoList.tsx`. Here's the change:

my-app/components/TodoList.tsx
```tsx
<<<<<<< SEARCH
  const handleDelete = (id: number) => {
    setTodos(todos.map((todo) => (todo.id === id ? { ...todo, completed: true } : todo)));
  };
=======
  const handleDelete = (id: number) => {
    setTodos(todos.filter((todo) => todo.id !== id));
  };
>>>>>>> REPLACE
```

In the current implementation, the `handleDelete` function is only marking the todo as completed by setting `completed` to `true`. This doesn't actually remove the todo from the list.

To fix this, we need to update the `handleDelete` function to remove the todo from the `todos` array using the `filter` method. The updated function will filter out the todo with the matching `id`, effectively removing it from the list.

With this change, pressing the delete key after tabbing to a todo item will remove the todo from the list.