aider: Added focus styles to "Clear All Todos" button for better visibility.
# Aider chat conversation:

USER: OK, it is included in the tabbing flow but it is not obvious. It doesn't have any highlighting to show that it can be selected.

ASSISTANT: To make the "Clear All Todos" button more obvious when it is focused, we can add some CSS styles to highlight it. We will add a `focus` style to the button.

Here are the changes:

my-app/app/page.tsx
```typescript
<<<<<<< SEARCH
            className="text-red-500 hover:text-red-700 focus:outline-none"
=======
            className="text-red-500 hover:text-red-700 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2"
>>>>>>> REPLACE
```

This change will add a ring around the button when it is focused, making it more obvious that it can be selected.